Add authentication

+++ graphql/resolvers/booking.js
@@ -6,9 +6,13 @@ const {
 } = require('../../helpers/resolverTransformers')
 
 module.exports = {
-    bookings: async () => {
+    bookings: async (args, req) => {
+        if (!req.isAuthenticated) {
+            throw new Error('Unauthenticated')
+        }
         try {
             const bookings = await Booking.find()
+
             return bookings.map(booking => {
                 return bookingTransformer(booking)
             })
@@ -16,22 +20,36 @@ module.exports = {
             throw err
         }
     },
-    bookEvent: async args => {
+    bookEvent: async (args, req) => {
+        if (!req.isAuthenticated) {
+            throw new Error('Unauthenticated')
+        }
+
         const fetchedEvent = await Event.findOne({ _id: args.eventId })
+
         const booking = new Booking({
-            user: '5e1a5140a2af7b2c54fdc44b',
+            user: req.userId,
             event: fetchedEvent
         })
+
         const result = await booking.save()
+
         return bookingTransformer(result)
     },
-    cancelBooking: async args => {
+    cancelBooking: async (args, req) => {
+        if (!req.isAuthenticated) {
+            throw new Error('Unauthenticated')
+        }
+
         try {
             const booking = await Booking.findById(args.bookingId).populate(
                 'event'
             )
+
             const event = eventTransformer(booking.event)
+
             await Booking.deleteOne({ _id: args.bookingId })
+
             return event
         } catch (err) {
             throw err
